Issue #15218: formatted input file for section 4.8.2.2 declared when needed to follow google style guide rules
